@@ -1,30 +1,42 @@
-cmake_minimum_required (VERSION 3.8)
+cmake_minimum_required (VERSION 3.7)
+cmake_policy(VERSION 3.7...3.20)
+
 project (cfd3d)
-set (CMAKE_CXX_STANDARD 11)
-option(USE_MPI "USE_MPI" OFF)
-option(USE_CUDA "USE_CUDA" OFF)
-option(USE_OPENCL "USE_OPENCL" OFF)
+set(CMAKE_CXX_STANDARD 17)
+
+option(USE_CUDA "USE_CUDA" ON)
+option(USE_OPENCL "USE_OPENCL" ON)
+option(USE_MPI "USE_MPI" ON)
 list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/CMake)
 #MESSAGE(STATUS "CMAKE_MODULE_PATH: " ${CMAKE_MODULE_PATH})
 
 file(GLOB_RECURSE SOURCES src/*.cpp src/*.c)
 file(GLOB_RECURSE CUDA_SOURCES src/*.cu)
+
 include_directories(src)
 
-if (USE_OPENCL)
-    add_definitions(-DUSE_OPENCL)
-else()
+if(MSVC)
+    set(CMAKE_CXX_FLAGS "-W3 /EHsc")
+elseif(CMAKE_COMPILER_IS_GNUCXX)
+    set(CMAKE_CXX_FLAGS "-Wall")
+elseif(CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
+    set(CMAKE_CXX_FLAGS "-Wall")
+endif()
+
+if(${USE_OPENCL})
+    find_package(OpenCL QUIET)
+endif()
+if (NOT ${USE_OPENCL} OR NOT OpenCL_FOUND)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Opencl/CfdSolverOpencl.cpp)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Opencl/ClInterface.cpp)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Opencl/SorSolverOpencl.cpp)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Opencl/UvwOpencl.cpp)
 endif()
 
-if (USE_MPI)
-    find_package(MPI)
-    include_directories(SYSTEM ${MPI_INCLUDE_PATH})
-    add_definitions(-DUSE_MPI)
-else()
+if(${USE_MPI})
+    find_package(MPI QUIET)
+endif()
+if (NOT ${USE_MPI} OR NOT MPI_FOUND)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Mpi/BoundaryValuesMpi.cpp)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Mpi/CfdSolverMpi.cpp)
     list(REMOVE_ITEM SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/CfdSolver/Mpi/MpiHelpers.cpp)
